# kaifa-be/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaifa API
  version: 1.0.0
  description: REST and WebSocket API for the Kaifa AI language learning platform.
servers:
  - url: https://api.kaifanesia.com/api
    description: Production API
  - url: http://localhost:8080/api
    description: Local API
...
security:
  - bearerAuth: []
paths:
...
  /learning/start:
    post:
      tags: [Learning]
      summary: Start a learning chat for a generated learning path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartLearningChatRequest"
      responses:
        "200":
          description: Opening tutor reply
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LearningChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /learning/speech/tts:
    post:
      tags: [Learning]
      summary: Convert learning text to speech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeechRequest"
      responses:
        "200":
          description: MP3 audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: invalid request body
...
    SpeechRequest:
      type: object
      required: [user_text]
      properties:
        user_text:
          type: string
          example: Hello, welcome to today's lesson.
...
    StartLearningChatRequest:
      type: object
      required: [learning_path_step_id]
      properties:
        learning_path_step_id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-step-1
...
    LearningChatResponse:
      type: object
      required: [type, reply]
      properties:
        type:
          type: string
          enum: [chat, guardrail, listening_paragraph, listening_question, listening_assessment, material_required]
          example: chat
        reply:
          type: string
        paragraph:
          type: string
          description: Present when type is listening_paragraph. The generated listening passage should be 100-150 words. Frontend can send this text to TTS.
        question_number:
          type: integer
          description: Present when type is listening_question.
          example: 1
        assessment:
          $ref: "#/components/schemas/ListeningAssessment"
        listening_state:
          $ref: "#/components/schemas/LearningListeningState"
    LearningListeningState:
      type: object
      properties:
        phase:
          type: string
          enum: [awaiting_ready, asking, completed]
          example: asking
        material_type:
          type: string
          enum: [listening, reading]
          example: listening
        paragraph:
          type: string
        questions_asked:
          type: integer
          example: 2
        answers:
          type: array
          items:
            type: string
    ListeningAssessment:
      type: object
      properties:
        score:
          type: integer
          example: 82
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
...